feat(portfolio): moved github repos logic to server route
